Fix the bug of multiboot application case
The boot module check fails in this test case since there is no module
provided for this test case in Grub configuration of CI test platform
and wrong check logic in the test case.

According to requirement "The physical platform shall guarantee that
each module load physical address is higher than 100000H.".
So we need to add parameter "--quirk-modules-after-kernel" in grub.
Because --quirk-modules-after-kernel needed for kernels which load
at relatively high address.

The check logic of the multiboot module address in the test code
is wrong and does not meet our requirements.
This patch is to fix the multiboot module address check logic.

Please Refer https://jira.devtools.intel.com/browse/ACRN-7172.

Signed-off-by: Hu Fenglin <fenglin.hu@intel.com>